feat: use new API Manager
