+++ .env.dev
@@ -1,3 +1,5 @@
+INSEE_USERNAME=
+INSEE_PASWWORD=
 INSEE_CLIENT_ID=
 INSEE_CLIENT_SECRET=
 INSEE_CLIENT_ID_FALLBACK=
