+++ .github/workflows/external-api-test.yml
@@ -12,6 +12,8 @@ on:
 # For critical env variable, declare it within the action.
 env:
   node-version: 18.x
+  INSEE_USERNAME: ${{ secrets.INSEE_USERNAME }}
+  INSEE_PASSWORD: ${{ secrets.INSEE_PASSWORD }}
   INSEE_CLIENT_ID: ${{ secrets.INSEE_CLIENT_ID }}
   INSEE_CLIENT_SECRET: ${{ secrets.INSEE_CLIENT_SECRET}}
   INSEE_CLIENT_ID_FALLBACK: ${{ secrets.INSEE_CLIENT_ID_FALLBACK }}
